requirements.txt

streamlit
google-generativeai
python-docx
PyMuPDF 
langdetect 
fpdf 
gTTS

